Transfer acf and related functions from plotting to stat_tools
